+++ hutool-core/src/main/java/cn/hutool/core/math/Calculator.java
@@ -155,11 +155,7 @@ public class Calculator {
 	 * @return 优先级
 	 */
 	public boolean compare(char cur, char peek) {// 如果是peek优先级高于cur，返回true，默认都是peek优先级要低
-		boolean result = false;
-		if (operatPriority[(peek) - 40] >= operatPriority[(cur) - 40]) {
-			result = true;
-		}
-		return result;
+		return operatPriority[(peek) - 40] >= operatPriority[(cur) - 40];
 	}
 
 	/**